requirements.txt
Django>=1.11.1
celery>=4.0
asgi_rabbitmq>=0.3
django-celery-results
channels>=1.1.3
asgi_redis
django-tinymce